CRUD operations finalized

+++ src/main/java/br/com/hotmart/java/services/AdressService.java
@@ -3,6 +3,7 @@ package br.com.hotmart.java.services;
 import br.com.hotmart.java.controllers.forms.AdressForm;
 import br.com.hotmart.java.controllers.vo.AdressVO;
 import br.com.hotmart.java.entities.Adress;
+import br.com.hotmart.java.exception.ResourceNotFoundException;
 import br.com.hotmart.java.repositories.AdressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,26 +21,49 @@ public class AdressService {
         return adressRepository.findAll().stream().map(AdressVO::new).collect(Collectors.toList());
     }
 
-    public void saveAdress(AdressForm form) {
+    public void save(AdressForm form) {
         Adress newAdress = new Adress(form);
         adressRepository.save(newAdress);
     }
 
     public AdressVO getById(Long id) {
-        return new AdressVO(findById(id));
+        Adress adress = adressRepository.findById(id).orElse(null);
+
+        if (adress == null) {
+            throw new ResourceNotFoundException("id", "Adress does not exist!");
+        }
+
+        return new AdressVO(adress);
     }
 
     public Adress findById(Long id) {
-        return adressRepository.findById(id).orElse(null);
+        Adress adress = adressRepository.findById(id).orElse(null);
+
+        if (adress == null) {
+            throw new ResourceNotFoundException("id", "Adress does not exist!");
+        }
+
+        return adress;
     }
 
     public void update(Long id, AdressForm form) {
-        Adress existingAdress = findById(id);
+        Adress existingAdress = adressRepository.findById(id).orElse(null);
+
+        if (existingAdress == null) {
+            throw new ResourceNotFoundException("id", "Adress does not exist!");
+        }
+
         existingAdress.update(form);
         adressRepository.save(existingAdress);
     }
 
     public void delete(Long id) {
+        Adress adress = adressRepository.findById(id).orElse(null);
+
+        if (adress == null) {
+            throw new ResourceNotFoundException("id", "Adress does not exist!");
+        }
+
         adressRepository.deleteById(id);
     }
 }
