+++ src/main/java/br/com/hotmart/java/services/DepartmentService.java
@@ -3,6 +3,7 @@ package br.com.hotmart.java.services;
 import br.com.hotmart.java.controllers.forms.DepartmentForm;
 import br.com.hotmart.java.controllers.vo.DepartmentVO;
 import br.com.hotmart.java.entities.Department;
+import br.com.hotmart.java.exception.ResourceNotFoundException;
 import br.com.hotmart.java.repositories.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,19 +27,44 @@ public class DepartmentService {
     }
 
     public DepartmentVO getById(Long id) {
-        try {
-            return new DepartmentVO(departmentRepository.findById(id).orElse(null));
-        } catch (NullPointerException e) {
-            return new DepartmentVO();
+        Department department = departmentRepository.findById(id).orElse(null);
+
+        if (department == null) {
+            throw new ResourceNotFoundException("id", "Department does not exist!");
         }
+
+        return new DepartmentVO(department);
+    }
+
+    public Department findById(Long id) {
+        Department department = departmentRepository.findById(id).orElse(null);
+
+        if (department == null) {
+            throw new ResourceNotFoundException("id", "Department does not exist!");
+        }
+
+        return department;
     }
 
     public void update(Long id, DepartmentForm form) {
         Department existingDepartment = departmentRepository.findById(id).orElse(null);
+
+        if (existingDepartment == null) {
+            throw new ResourceNotFoundException("id", "Department does not exist!");
+        }
+
         existingDepartment.update(form);
+        departmentRepository.save(existingDepartment);
     }
 
     public void delete(Long id) {
+        Department department = departmentRepository.findById(id).orElse(null);
+
+        if (department == null) {
+            throw new ResourceNotFoundException("id", "Department does not exist!");
+        }
+
         departmentRepository.deleteById(id);
     }
+
 }
