+++ src/main/java/br/com/hotmart/java/services/EmployeeService.java
@@ -7,7 +7,6 @@ import br.com.hotmart.java.entities.Employee;
 import br.com.hotmart.java.exception.ResourceNotFoundException;
 import br.com.hotmart.java.repositories.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +19,7 @@ public class EmployeeService {
     private EmployeeRepository employeeRepository;
 
     @Autowired
-    private AdressService adressService;
+    private ProjectService projectService;
 
     public List<EmployeeVO> getAll() {
         return employeeRepository.findAll().stream()
@@ -28,31 +27,67 @@ public class EmployeeService {
     }
 
     public void save(EmployeeForm form) {
-        Adress adress = adressService.findById(form.getAdressId());
         Employee newEmployee = new Employee(form);
-        newEmployee.setAdress(adress);
+
+        newEmployee.setAdress(new Adress(form.getAdressForm()));
+
+        if (form.getSupervisorId() != null)
+            newEmployee.setSupervisor(this.findById(form.getSupervisorId()));
+
+        newEmployee.setProjects(projectService.findAllById(form.getProjectId()));
 
         employeeRepository.save(newEmployee);
     }
 
     public EmployeeVO getById(Long id) {
-
         Employee employee = employeeRepository.findById(id).orElse(null);
 
         if (employee != null) {
             return new EmployeeVO(employee);
         }
 
-        throw new ResourceNotFoundException("id", "Employee not found.");
+        throw new ResourceNotFoundException("id", "Employee does not exist!");
+    }
+
+    public Employee findById(Long id) {
+        Employee employee = employeeRepository.findById(id).orElse(null);
+
+        if (employee == null) {
+            throw new ResourceNotFoundException("id", "Employee does not exist!");
+        }
+
+        return employee;
     }
 
     public void update(Long id, EmployeeForm form) {
-        Employee existingEmployee = employeeRepository.findById(id).orElse(new Employee());
+        Employee existingEmployee = employeeRepository.findById(id).orElse(null);
+
+        if (existingEmployee == null) {
+            throw new ResourceNotFoundException("id", "Employee does not exist!");
+        }
+
         existingEmployee.update(form);
         employeeRepository.save(existingEmployee);
     }
 
+    public void updateSupervisor(Long employeeId, Long supervisorId) {
+        Employee existingEmployee = employeeRepository.findById(employeeId).orElse(null);
+
+        if (existingEmployee == null) {
+            throw new ResourceNotFoundException("id", "Employee does not exist!");
+        }
+
+        existingEmployee.setSupervisor(this.findById(supervisorId));
+        employeeRepository.save(existingEmployee);
+    }
+
     public void delete(Long id) {
+        Employee employee = employeeRepository.findById(id).orElse(null);
+
+        if (employee == null) {
+            throw new ResourceNotFoundException("id", "Employee does not exist!");
+        }
+
         employeeRepository.deleteById(id);
     }
 }
